rename: 모달 폴더 구분
